Create Electricity bill.cpp
Calculate the amount of money to be paid based on certain scenarios.

After the announcement by the government of its intention to pay 50% of the electricity bill of all categories of consumer, the board of Electricity Company of Ghana (ECG) were in a dilemma as to whether to adjust the old billing software or design a new one. It was decided that a new software should be designed because the old one was designed using FORTRAN and there are no FORTRAN programmers around. Each time they require to make changes to the software, they had to invite the designers from US, this is quite expensive than building a new one locally. You have won a contract to design the software using c++. On a trial basis you are to design a console application that will

a) request for i. customer name ii quantity of electricity used iii categorize customers into

domestic when at most 150 KW units is used within
industrial when more than 150 kw but less than or equal to 350 KW is used
commercial when more than 350 KW is used (8 marks)CR
b) if the cost per KW for domestic, industrial and commercial are respectively Ghc 1.2, 2.3 and 3.5 respectively. Your program should

compute the monthly bill based on category.
Add 1% of the monthly bill for rural electrification to get the total bill
reduce the total bill by 50% (8 marks) EV
c) Display the following results

customer’s name
quantity of water consumed and category
1% rural electrification charge and total amount for the month
50% government subsidy
balance to be paid (4 marks) CR NB: use the appropriate data type for all the variables